added "me" endoint and done some refactors

+++ src/auth/auth.service.ts
@@ -7,11 +7,12 @@ import { CreateUserDto } from "src/users/dto/create-user.dto";
 import { UsersService } from "src/users/users.service";
 import * as bcrypt from "bcrypt";
 import { TokensService } from "src/tokens/tokens.service";
-import { Response } from "express";
+import { CookieOptions, Response } from "express";
 import { ConfigService } from "@nestjs/config";
 import { JwtPayload } from "../classes/jwt-payload";
-import { AppRequest } from "../interfaces/app-request";
+import { AppRequest } from "../interfaces/app-request.interface";
 import { ExceptionMessages } from "../constants/exception-messages";
+import { LoginDto } from "./dto/login.dto";
 
 @Injectable()
 export class AuthService {
@@ -21,7 +22,11 @@ export class AuthService {
     private configService: ConfigService
   ) {}
 
-  async registration(req: AppRequest, createUserDto: CreateUserDto) {
+  async registration(
+    req: AppRequest,
+    res: Response,
+    createUserDto: CreateUserDto
+  ) {
     const user = await this.usersService.getUserByLogin(createUserDto.login);
 
     if (user) {
@@ -41,9 +46,12 @@ export class AuthService {
 
     await this.tokensService.createToken({
       refreshToken: tokens.refreshToken,
+      rememberMe: true,
       userAgent: req.headers["user-agent"],
       userId: newUser.id,
     });
+
+    this.setRefreshCookie(res, tokens.refreshToken);
 
     return {
       user: {
@@ -53,7 +61,7 @@ export class AuthService {
     };
   }
 
-  async login(req: AppRequest) {
+  async login(req: AppRequest, res: Response, loginDto: LoginDto) {
     const userAgent = req.headers["user-agent"];
     const payload = req.user;
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -62,19 +70,27 @@ export class AuthService {
     );
     const tokens = this.tokensService.generateTokens(user);
     const sessions = await this.tokensService.findTokensByUserId(user.id);
+    const tokenOptions = {
+      refreshToken: tokens.refreshToken,
+      expires: loginDto.rememberMe
+        ? new Date(Date.now() + this.configService.get("refreshTokenAliveTime"))
+        : null,
+      rememberMe: loginDto.rememberMe,
+      userAgent,
+    };
 
     if (sessions.length < this.configService.get("maxSessionsPerAcc")) {
       await this.tokensService.createToken({
-        refreshToken: tokens.refreshToken,
-        userAgent: userAgent,
+        ...tokenOptions,
         userId: user.id,
       });
     } else {
       await this.tokensService.updateFirstByUserId(user.id, {
-        refreshToken: tokens.refreshToken,
-        userAgent,
+        ...tokenOptions,
       });
     }
+
+    this.setRefreshCookie(res, tokens.refreshToken, loginDto.rememberMe);
 
     return {
       user: {
@@ -84,22 +100,59 @@ export class AuthService {
     };
   }
 
-  async refresh(req: AppRequest) {
-    const oldToken = req.cookies["refreshToken"];
+  async me(refreshToken: string) {
+    if (!refreshToken) {
+      throw new UnauthorizedException(ExceptionMessages.Unauthorized);
+    }
+
+    const token = await this.tokensService.findToken(refreshToken);
+
+    if (token) {
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      const { password, ...user } = await this.usersService.getUserById(
+        token.userId
+      );
+      return user;
+    }
+
+    throw new UnauthorizedException(ExceptionMessages.Unauthorized);
+  }
+
+  async refresh(req: AppRequest, res: Response) {
+    const oldRefreshToken = req.cookies["refreshToken"];
+
+    if (!oldRefreshToken) {
+      throw new UnauthorizedException(ExceptionMessages.Unauthorized);
+    }
+
     const userAgent = req.headers["user-agent"];
 
-    if (oldToken) {
-      const session = await this.tokensService.findToken(oldToken);
+    if (oldRefreshToken) {
+      const session = await this.tokensService.findToken(oldRefreshToken);
 
       if (session) {
         const user = await this.usersService.getUserById(session.userId);
         const payload = new JwtPayload(user);
         const tokens = this.tokensService.generateTokens(payload);
 
-        await this.tokensService.updateToken(oldToken, {
-          refreshToken: tokens.refreshToken,
-          userAgent,
-        });
+        const updatedToken = await this.tokensService.updateToken(
+          oldRefreshToken,
+          {
+            refreshToken: tokens.refreshToken,
+            expires: session.rememberMe
+              ? new Date(
+                  Date.now() + this.configService.get("refreshTokenAliveTime")
+                )
+              : null,
+            userAgent,
+          }
+        );
+
+        this.setRefreshCookie(
+          res,
+          tokens.refreshToken,
+          updatedToken.rememberMe
+        );
 
         return tokens;
       }
@@ -108,15 +161,19 @@ export class AuthService {
     throw new UnauthorizedException(ExceptionMessages.Unauthorized);
   }
 
-  async logout(refreshToken: string) {
+  async logout(req: AppRequest, res: Response) {
+    const refreshToken = req.cookies["refreshToken"];
+
+    if (!refreshToken) {
+      throw new UnauthorizedException(ExceptionMessages.Unauthorized);
+    }
+
     const session = await this.tokensService.findToken(refreshToken);
 
     if (session) {
       await this.tokensService.deleteToken(refreshToken);
-
-      return {
-        refreshToken: session.refreshToken,
-      };
+      this.deleteRefreshCookie(res);
+      return;
     }
 
     throw new UnauthorizedException(ExceptionMessages.Unauthorized);
@@ -138,11 +195,25 @@ export class AuthService {
     return null;
   }
 
-  setCookie(res: Response, value: string) {
+  private setRefreshCookie(res: Response, value: string, rememberMe = true) {
     res.cookie("refreshToken", value, {
-      httpOnly: true,
-      maxAge: this.configService.get("cookieAliveTime") * 24 * 60 * 60 * 1000,
-      path: "auth",
+      ...refreshCookieOptions,
+      expires: rememberMe
+        ? new Date(Date.now() + this.configService.get("refreshTokenAliveTime"))
+        : null,
+    });
+  }
+
+  private deleteRefreshCookie(res: Response) {
+    res.clearCookie("refreshToken", {
+      ...refreshCookieOptions,
+      maxAge: 0,
     });
   }
 }
+
+const refreshCookieOptions: CookieOptions = {
+  httpOnly: true,
+  sameSite: "lax",
+  path: "auth",
+};
